+++ app/src/main/java/woowacourse/shopping/data/service/okhttp/product/ProductServiceImpl.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.data.service.product
+package woowacourse.shopping.data.service.okhttp.product
 
 import okhttp3.RequestBody.Companion.toRequestBody
 import woowacourse.shopping.ShoppingApplication.Companion.pref
 import woowacourse.shopping.data.model.Product
-import woowacourse.shopping.data.service.cart.ProductId
+import woowacourse.shopping.data.service.okhttp.cart.ProductId
 import woowacourse.shopping.data.util.convertJsonToProduct
 import woowacourse.shopping.data.util.convertJsonToProducts
 import woowacourse.shopping.data.util.okhttp.Header.Companion.JSON_MEDIA_TYPE
